requirements.txt
mcp>=1.0.0
requests
python-dotenv
streamlit
plotly
